land b_md onto HEAD:  - LRU counters were broken, causing constant lock purge (433, 432)  - the first part of the fixes for the ldlm_handle2lock race (419)  - cleanup unused ll_inode_info field (442)  - keep ioctl semaphore from serializing test_getattr/brw

+++ lustre/include/linux/obd.h
@@ -273,7 +273,7 @@ struct obd_device {
 };
 
 struct obd_ops {
-        int (*o_iocontrol)(long cmd, struct lustre_handle *, int len,
+        int (*o_iocontrol)(unsigned int cmd, struct lustre_handle *, int len,
                            void *karg, void *uarg);
         int (*o_get_info)(struct lustre_handle *, obd_count keylen, void *key,
                           obd_count *vallen, void **val);
